% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyorsad.R
\name{happyorsad}
\alias{happyorsad}
\title{Sentiment scores a string}
\usage{
happyorsad(string, language = "da")
}
\arguments{
\item{string:}{a text string}

\item{language:}{"en" for english, "da" for danish, "emoticon" for emoticons.}
}
\value{
The function returns the sum of sentiment.
}
\description{
happyorsad sentiment-scores a string of text. It uses the AFINN sentiment scoring. It assigns a value to each recognized word or emoticon in the string and returns the sum.
}
\examples{
# Examples of sentiment scoring

library(happyorsad)

# Score danish words
string_da <- 'Hvis ikke det er det mest afskyelige elendige flueknepperi...'
happyorsad(string_da, "da")

# Score english words
string_en <- 'This is utterly excellent!'
happyorsad(string_en, "en")

# Score emoticons
string_emoticon <- 'I saw that yesterday :)'
happyorsad(string_emoticon, "emoticon")
}

