% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentida.R
\name{sentida}
\alias{sentida}
\title{Sentiment scores a string}
\usage{
sentida(string, output = "total")
}
\arguments{
\item{string}{a text string}

\item{output}{"total" for total score, "mean" for mean score}

\item{x}{value to match}

\item{table}{object to match against}
}
\value{
The function returns the sum of sentiment.
}
\description{
Sentida sentiment-scores a string of text.
}
\examples{

sentida("Gud bevare Danmark")



}
